request_utility의 일부 함수를 utility 모듈로 이동
